+++ pkg/storage/stores/indexshipper/uploads/index_set_test.go
@@ -36,10 +36,12 @@ func TestIndexSet_Add(t *testing.T) {
 
 			// see if we can find all the added indexes in the table.
 			indexesFound := map[string]*mockIndex{}
-			err = indexSet.ForEach(func(_ bool, index index.Index) error {
+			doneChan := make(chan struct{})
+			err = indexSet.ForEach(context.Background(), doneChan, func(_ bool, index index.Index) error {
 				indexesFound[index.Path()] = index.(*mockIndex)
 				return nil
 			})
+			close(doneChan)
 			require.NoError(t, err)
 
 			require.Equal(t, testIndexes, indexesFound)
@@ -107,10 +109,12 @@ func TestIndexSet_Cleanup(t *testing.T) {
 
 			// all the indexes should be retained since they were just uploaded
 			indexesFound := map[string]*mockIndex{}
-			err = idxSet.ForEach(func(_ bool, index index.Index) error {
+			doneChan := make(chan struct{})
+			err = idxSet.ForEach(context.Background(), doneChan, func(_ bool, index index.Index) error {
 				indexesFound[index.Path()] = index.(*mockIndex)
 				return nil
 			})
+			close(doneChan)
 			require.NoError(t, err)
 
 			require.Equal(t, testIndexes, indexesFound)
@@ -131,10 +135,12 @@ func TestIndexSet_Cleanup(t *testing.T) {
 
 			// get all the indexes that are retained
 			indexesFound = map[string]*mockIndex{}
-			err = idxSet.ForEach(func(_ bool, index index.Index) error {
+			doneChan = make(chan struct{})
+			err = idxSet.ForEach(context.Background(), doneChan, func(_ bool, index index.Index) error {
 				indexesFound[index.Path()] = index.(*mockIndex)
 				return nil
 			})
+			close(doneChan)
 			require.NoError(t, err)
 
 			// we should have only the indexes whose upload time was not changed above